Major UI/Functionality update
Details page will now be hidden untill user click on a tool/part, to increase working soace in case of search only usage.
Adding a small table displaying which parts currently is in the clicked tool
Format all the tool details into a smaller table beside that.
Move addItem to the main tabs, since it make no sense to click a tool/part first to allow adding new thing to the json